Refactor Form from create to use it for edit
